requirements.txt
aniso8601==8.0.0
asn1crypto==1.3.0
attrs==19.3.0
Babel==2.8.0
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.6.1
cymem==2.0.3
Delorean==1.0.0
en-core-web-sm==2.2.5
Faker==1.0.7
Flask==1.0.2
flask-restx==0.2.0
Flask-SQLAlchemy==2.3.1
freezegun==0.1.11
gunicorn==20.0.4
humanize==2.4.0
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
more-itertools==8.2.0
murmurhash==1.0.2
numpy==1.18.4
packaging==20.3
parse==1.12.0
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
py==1.8.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.1
pytest-flask==0.14.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
six==1.14.0
spacy==2.2.4
SQLAlchemy==1.3.16
srsly==1.0.2
text-unidecode==1.2
thinc==7.4.0
tqdm==4.46.0
tzlocal==2.0.0
urllib3==1.25.9
wasabi==0.6.0
wcwidth==0.1.9
Werkzeug==0.16.0
zipp==3.1.0
